Set sub-destructible timeout and inherit parent's.
Managing the timing of timeouts is becoming a major part of working with
Destructible. There was no way to set the timeout of ephemeral
sub-destructibles and the default of one second is not the size that
fits all. This new implementation will propagate the parent's timeout so
that ephemerals will have the same timeout as the root Destructible. You
can set the timeout of durables and ephemerals (and forgivables) so that
all their children can have a different deadline. Note that ephemerals
run on a different schedule than the parent so they could be allowed to
take longer, but that doesn't make sense when the parent may tell
everyone to shutdown.

Actually, I'm not quite sure what to do when shutdown means shutting
down thousands of connected sockets, so we have further development
here. Not sure if the shutdowns fire in parallel or sequentially, but
they probably run in parallel since destroy is synchronous and a
sub-destructible is not.