Remove `errored` argument to destructor.
Not expecting it in the down stream modules. It is causing problems when
a destructor is created using `Function.prototype.bind` on variadic
functions that do something special when additional arguments are
provided. Specifically, `Descendant.down` treats expects the argument to
be a `Socket`, but it is getting a `boolean`.

Don't need to pass the flag. It is used when creating ephemeral
sub-destructible to determine if the sub-destructible `errored`, but we
can just check to see whether or not its `_error` array is empty.

Closes #157.