Remove `cause`.
Appears to be a legacy. I've never inspected the errors. Interrupt is
doing some form of nesting of causes and Rescue provides a way to match
against them. The nesting of errors is not part of Destructible.
Destructible isn't even there as far as that goes. It disappears as
strands disappear. It is a chimera. A verb, not a noun. A presence, not
an object.

Closes #158.