Remove user scrammable.
Was trying to sort out what to do with a a `Destructible` component that
wasn't supposed to destroy itself using `destruct`, like a socket
channel that should really be told by the application to send an
end-of-stream. Found the answeres in Conduit and recorded them in the
diary here.

See #146.