Notes on scrammable, minor code tidy.
Thought I was going to make `_scrams` a list of `_children`, but see
that I'm piggybacking timer shutdown with a push of a stop timer to the
end of that list. Upon reflection, it still makes sense after
`async`/`await` conversion, but it was the obvious choice with
error-first callbacks when `_scram` was synchronous.

Closes #146.