Errored propagation through destruct channel.
We can't use the length of the errors array as an indicator of error
state in response to the destruction of a child when the child may have
destructed because of an error in its child. The child will check the
errors array of its child, the grandchild, and destroy itself, but it
won't have an error for its own errors array yet when the parent checks
it.

Closes #163.