Implement `drain`.
Add a function that will wait for all the ephemeral strands to exit.
This has become necessary now that we are strict about adding ephemeral
strands after destruction, or rather, now that we're thinking about work
at destruction at all. User needs some way to wait for the work they've
put in the background to finish, and queue new final work that may or
may not be part of destruction, per se. Allows destructible to be used
as something of a work queue. Don't abuse it.

Closes #212.