Always fire working.
A lot of complexity was added to `Destructible` so that I wouldn't
traverse this linked list until destruction. It's silly. If the
`working()` method is called before every significant `await` call it
will be negligable compared to the `await` call. Calling it before
reading a file or while waiting for a socket to close is not expensive
during normal operations. I've introduced a child list in any case which
I can remove or maintain it explicitly for the purpose of avoiding this
linked-list traversal.

That is, if the `Destructible` is a child destructible, the working
call...

Working can be an object shared by the tree. Ouch.