Upgrade `interrupt` to `11.0.0-alpha.3`.
Interrupt is redesigned. Key is that there is no longer a context array
argument so we provide error context by wrapping an error from a strand
in a `Destructible.Error` if it isn't already a `Destructible.Error`.

We also added a stack tracer poker function which may or may not be
useful. The context exception is generated through the poker function so
that the user application file and line number appears in the stack
trace.

https://github.com/bigeasy/interrupt/releases/tag/v11.0.0-alpha.2