Upgrade `interrupt` to `11.0.0-alpha.11`.
Upgrade Interrupt to 11.0.0-alpha.11 which sorts out the trace function
which may still be useful to record the entry point of a destructible.

Finding that with Node.js 14 and some understanding of how `async`
functions work you can get meaningful stack traces out of ordinary
`Error`s, the nested stack traces we offer provide more meaning. Need to
describe how stack traces get truncated after passing an `await`, but
it's not as bad as it was with error-first callbacks.

Going to find a way to make a trace function available to all
destructibles. Not a property of the destructible itself because an
additional tracing does not need to merge the `Destructible`'s trace. It
will be there at the level of the destructible in the nested. You'll
have to learn to assemble it in your mind when you read the trace.

https://github.com/bigeasy/interrupt/releases/tag/v11.0.0-alpha.11