Fix initialization, rescue, exceptional.
Exceptional was throwing `DESTROYED` when it really should thrown
something indicating an error occured in the strand. The user said they
wanted to wait on the strand and if there where any errors to throw
them.

`rescue` now sets up a strand directly rather than calling a builder.
This allows it to muck around with the `vargs` so we can have a trace
function for `rescue` as well.

Now initialation works as expected. Fixed the unit test. Introduced
rescue to assert properties of exceptions in the unit test.