Thoughts on `recoverable`.
How do you recover an ephemeral? Moreover, the problem of
initialization. How do you ensure that a strand will be a good citizen
within a destructible tree? `Destructible.recoverable()` won't be
happening, but there are probably some ways to describe what we want to
accomplish now that we've created a vocabulary.