Moved preamble from old Docco.
Removed the preamble, actually, and rewrote it in `README.in.md`
starting with a description of strands that will probably work out fine,
and then created an example of strands in programs, but it's somewhat
difficult because trivial `async` programs are trivial.

Advocacy is going to be weak unless we use sweaping examples.