Do not switch shutdown if waits are empty.
Hanging when a timed ephemeral shutdown switched to a durable shutdown.
The durable shutdown is on a timer. If the parent shuts down the
destructable is marked durable and timer is canceled. If the
destructible did empty while it was on the time it should not enter
a durable shutdown. It should instead finish the current shutdown.